refactor(crepe): rename providers/ subpath to llm-providers/
`providers` is too vague in the milkdown context — could be confused
with editor / kit providers. `llm-providers` is unambiguous.

- Source: src/providers/ → src/llm-providers/
- Subpath: @milkdown/crepe/providers/{openai,anthropic} →
  @milkdown/crepe/llm-providers/{openai,anthropic}
- rollup + package.json exports updated accordingly
